RQ-216 Moved favicon to separate component
