+++ app/src/components/misc/Favicon/favicon.scss
@@ -1,0 +1,4 @@
+.ant-avatar.rq-favicon {
+  height: 16px;
+  width: 16px;
+}
