+++ app/src/views/features/api-client/client-view/request/requestTabs.scss
@@ -50,7 +50,7 @@
 
     .key,
     .name {
-      width: 200px;
+      width: 250px;
     }
 
     .value {
